requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements/prod.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
click==8.1.7
    # via uvicorn
exceptiongroup==1.2.2
    # via anyio
fastapi==0.112.1
    # via -r requirements/prod.in
gunicorn==23.0.0
    # via -r requirements/prod.in
h11==0.14.0
    # via uvicorn
idna==3.7
    # via anyio
packaging==24.1
    # via gunicorn
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.4.0
    # via -r requirements/prod.in
python-dotenv==1.0.1
    # via
    #   -r requirements/prod.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
starlette==0.38.2
    # via fastapi
tenacity==9.0.0
    # via -r requirements/prod.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
uvicorn==0.30.6
    # via -r requirements/prod.in
